requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
psycopg2-binary
